+++ FE/src/hooks/useFetch.js
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useReducer } from "react";
-import { sendAuthRequest } from "@/service/api";
+import { sendAuthRequest } from "@/service/authenticationAPI";
 import { BASE_URL } from "@/constants/constants";
 
 const actionType = {
