+++ FE/src/pages/Map/DriverMap.jsx
@@ -11,7 +11,7 @@ import { getCurrentLocation } from "@/utils/coords";
 import { getAccessToken } from "@/utils/auth";
 import { DriverBottomSheet } from "@/components/Bottomsheet/Bottomsheet";
 import Car from "@/assets/car.svg";
-import { getKidInfo } from "@/service/api";
+import { getKidInfo } from "@/service/childrenAPI";
 
 export default function DriverMap() {
     const ACCESS_TOKEN = getAccessToken();
