+++ FE/src/pages/Menu/Menu.jsx
@@ -1,5 +1,5 @@
 import { useLoaderData } from "react-router-dom";
-import { getKidInfo } from "@/service/api";
+import { getKidInfo } from "@/service/childrenAPI";
 import styles from "./Menu.module.scss";
 import { MenuButton } from "./MenuButton/MenuButton";
 import iDropGreen from "@/assets/iDropGreen.svg";
