+++ FE/src/pages/Parents/History/History.jsx
@@ -1,7 +1,7 @@
 import { Header } from "@/components/Header/Header";
 import styles from "./History.module.scss";
 import { HistoryItem } from "./HistoryItem/HistoryItem";
-import { getSubscriptionHistoryList } from "./api";
+import { getSubscriptionHistoryList } from "./parentsAPI";
 import { useLoaderData } from "react-router-dom";
 
 export default function History() {
