+++ FE/src/pages/Parents/Subscription/DriverDetail/DriverDetail.jsx
@@ -1,7 +1,7 @@
 import { useLoaderData, useLocation, useNavigate } from "react-router-dom";
 import { Footer } from "@/components/Footer/Footer";
 import { Header } from "@/components/Header/Header";
-import { getDriverDetail, postSubscribe } from "@/service/api";
+import { getDriverDetail, postSubscribe } from "@/service/parentsAPI";
 import styles from "./DriverDetail.module.scss";
 
 export default function DriverDetail() {
